feat: add very basic load generator for javascript SDK
This commit adds a very basic load generator that can
be used to experiment with the performance of the
momento javascript client on a single nodejs process.
It includes configurable settings for number of
concurrent requests, payload size, and request timeout.

Includes disclaimers to highlight that on nodejs, it's
easy to become CPU-bound, so that is often the bottleneck
for performance.

In future commits we will make the error handling / reporting
more user friendly, along with commits to the javascript
SDK itself to improve the UX for error handling.