feat: basic python load generator, fix tps measurements for nodejs (#98)
* feat: basic python load generator, fix tps measurements for nodejs

This commit introduces a very basic load generator tool for the
python SDK.  It measures throughput and latencies for requests
from a single python process.

This is basically a direct port of the earlier one that we wrote
for the JavaScript SDK.

Also included in this commit: I realized that my tps counts for
the javascript load generator were incorrect; I was counting
an "operation" (1 set + 1 get) as the unit of measurement, so
the tps I was reporting was actually only half of the real
value.  This commit fixes the computation for the JS SDK, and
makes sure the python version uses the correct logic.