fix: handle throttling errors in python loadgen
Prior to this commit, if a throttling error was encountered while
running the python load generator, it was not caught, and the
tool would crash.  This commit updates the code to handle the
exception; now, instead of crashing, we simply track the number
of successful requests vs. the number of throttled requests, and
we include a warning log message that lets the user know they
can reach out to us for a limit increase.