feat: remove dotnet examples, change README to link to SDKs
This commit removes the dotnet examples, which are being moved
to the SDK repo in another PR. Also removes related dependabot/CI
configs; these are also moved over to C# repo.

We also tweak the README to turn this repo into more of a splash
page to link to the other repos for the SDKs instead of hosting
the examples here.

Other examples aren't deleted yet, we will keep them around
until they're moved over to their repos.